requirements.txt
Django==4.2.1
django_paypal==2.0
